+++ src/test/java/ee/sk/smartid/LinkedNotificationSignatureSessionRequestBuilderTest.java
@@ -1,0 +1,223 @@
+package ee.sk.smartid;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import java.util.List;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import ee.sk.smartid.exception.UnprocessableSmartIdResponseException;
+import ee.sk.smartid.exception.permanent.SmartIdRequestSetupException;
+import ee.sk.smartid.rest.SmartIdConnector;
+import ee.sk.smartid.rest.dao.DeviceLinkInteraction;
+import ee.sk.smartid.rest.dao.LinkedSignatureSessionRequest;
+import ee.sk.smartid.rest.dao.LinkedSignatureSessionResponse;
+
+class LinkedNotificationSignatureSessionRequestBuilderTest {
+
+    private static final String DOCUMENT_NUMBER = "PNOEE-12345678901-MOCK-Q";
+    private SmartIdConnector connector;
+
+    @BeforeEach
+    void setUp() {
+        connector = mock(SmartIdConnector.class);
+    }
+
+    @Test
+    void initSignatureSession_ok() {
+        LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withDocumentNumber(DOCUMENT_NUMBER)
+                .withSignatureAlgorithm(SignatureAlgorithm.RSASSA_PSS)
+                .withSignableData(new SignableData("Test data".getBytes()))
+                .withLinkedSessionID("10000000-0000-0000-0000-000000000000")
+                .withInteractions(List.of(DeviceLinkInteraction.displayTextAndPIN("Sign?")));
+        when(connector.initLinkedNotificationSignature(any(LinkedSignatureSessionRequest.class), eq(DOCUMENT_NUMBER))).thenReturn(new LinkedSignatureSessionResponse("20000000-0000-0000-0000-000000000000"));
+
+        LinkedSignatureSessionResponse response = builder.initSignatureSession();
+        assertEquals("20000000-0000-0000-0000-000000000000", response.sessionID());
+    }
+
+    @ParameterizedTest
+    @EnumSource(CertificateLevel.class)
+    void initSignatureSession_withDifferentCertificateLevels_ok(CertificateLevel certificateLevel) {
+        LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withCertificateLevel(certificateLevel)
+                .withDocumentNumber(DOCUMENT_NUMBER)
+                .withSignableData(new SignableData("Test data".getBytes()))
+                .withLinkedSessionID("10000000-0000-0000-0000-000000000000")
+                .withInteractions(List.of(DeviceLinkInteraction.displayTextAndPIN("Sign?")));
+        when(connector.initLinkedNotificationSignature(any(LinkedSignatureSessionRequest.class), eq(DOCUMENT_NUMBER))).thenReturn(new LinkedSignatureSessionResponse("20000000-0000-0000-0000-000000000000"));
+
+        LinkedSignatureSessionResponse response = builder.initSignatureSession();
+        assertEquals("20000000-0000-0000-0000-000000000000", response.sessionID());
+    }
+
+    @Nested
+    class ValidateRequestParameters {
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        void initSignatureSession_relyingPartyUUIDIsEmpty_throwException(String relyingPartyUUID) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID(relyingPartyUUID);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class, builder::initSignatureSession);
+            assertEquals("Value for 'relyingPartyUUID' cannot be empty", ex.getMessage());
+        }
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        void initSignatureSession_relyingPartyNameIsEmpty_throwException(String relyingPartyName) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName(relyingPartyName);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class, builder::initSignatureSession);
+            assertEquals("Value for 'relyingPartyName' cannot be empty", ex.getMessage());
+        }
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        void initSignatureSession_documentNumberIsEmpty_throwException(String documentNumber) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(documentNumber);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class, builder::initSignatureSession);
+            assertEquals("Value for 'documentNumber' cannot be empty", ex.getMessage());
+        }
+
+        @Test
+        void initSignatureSession_signableDataOrSignableHashNotProvided_throwException() {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER)
+                    .withSignableData(null)
+                    .withSignableHash(null);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class, builder::initSignatureSession);
+            assertEquals("Value for 'digestInput' must be set with SignableData or with SignableHash", ex.getMessage());
+        }
+
+        @Test
+        void initSignatureSession_signableDataAlreadyUsedForSettingDigest_throwException() {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withSignableData(new SignableData("Test data".getBytes()))
+                            .withSignableHash(new SignableHash(DigestCalculator.calculateDigest("Test data".getBytes(), HashAlgorithm.SHA_512))));
+            assertEquals("Value for 'digestInput' has been already set with SignableData", ex.getMessage());
+        }
+
+        @Test
+        void initSignatureSession_signableHashAlreadyUsedForSettingDigest_throwException() {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER);
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withSignableHash(new SignableHash(DigestCalculator.calculateDigest("Test data".getBytes(), HashAlgorithm.SHA_512)))
+                            .withSignableData(new SignableData("Test data".getBytes())));
+            assertEquals("Value for 'digestInput' has been already set with SignableHash", ex.getMessage());
+        }
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        void initSignatureSession_linkedSessionIDIsEmpty_throwException(String linkedSessionID) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER)
+                    .withSignableData(new SignableData("Test data".getBytes()));
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withLinkedSessionID(linkedSessionID)
+                            .initSignatureSession());
+            assertEquals("Value for 'linkedSessionID' cannot be empty", ex.getMessage());
+        }
+
+        @ParameterizedTest
+        @ValueSource(strings = {"1234567890123456789012345678901", ""})
+        void initSignatureSession_nonceWithIncorrectLengthProvided_throwException(String nonce) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER)
+                    .withSignableData(new SignableData("Test data".getBytes()))
+                    .withLinkedSessionID("10000000-0000-0000-0000-000000000000")
+                    .withInteractions(List.of(DeviceLinkInteraction.displayTextAndPIN("Sign?")));
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withNonce(nonce)
+                            .initSignatureSession());
+            assertEquals("Value for 'nonce' must be 1-30 characters long", ex.getMessage());
+        }
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        void initSignatureSession_interactionsInEmpty_throwException(List<DeviceLinkInteraction> interactions) {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER)
+                    .withSignableData(new SignableData("Test data".getBytes()))
+                    .withLinkedSessionID("10000000-0000-0000-0000-000000000000");
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withInteractions(interactions)
+                            .initSignatureSession());
+            assertEquals("Value for 'interactions' cannot be empty", ex.getMessage());
+        }
+
+        @Test
+        void initSignatureSession_interactionsContainDuplicates_throwException() {
+            LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                    .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                    .withRelyingPartyName("DEMO")
+                    .withDocumentNumber(DOCUMENT_NUMBER)
+                    .withSignableData(new SignableData("Test data".getBytes()))
+                    .withLinkedSessionID("10000000-0000-0000-0000-000000000000");
+
+            var ex = assertThrows(SmartIdRequestSetupException.class,
+                    () -> builder.withInteractions(List.of(DeviceLinkInteraction.displayTextAndPIN("Sign?"), DeviceLinkInteraction.displayTextAndPIN("Sign again?")))
+                            .initSignatureSession());
+            assertEquals("Value for 'interactions' cannot contain duplicate types", ex.getMessage());
+        }
+    }
+
+    @Test
+    void initSignatureSession_sessionIDMissingFromResponse_throwException() {
+        LinkedNotificationSignatureSessionRequestBuilder builder = new LinkedNotificationSignatureSessionRequestBuilder(connector)
+                .withRelyingPartyUUID("00000000-0000-0000-0000-000000000000")
+                .withRelyingPartyName("DEMO")
+                .withDocumentNumber(DOCUMENT_NUMBER)
+                .withSignableData(new SignableData("Test data".getBytes()))
+                .withLinkedSessionID("10000000-0000-0000-0000-000000000000")
+                .withInteractions(List.of(DeviceLinkInteraction.displayTextAndPIN("Sign?")));
+        when(connector.initLinkedNotificationSignature(any(LinkedSignatureSessionRequest.class), eq(DOCUMENT_NUMBER))).thenReturn(new LinkedSignatureSessionResponse(null));
+
+        var ex = assertThrows(UnprocessableSmartIdResponseException.class, builder::initSignatureSession);
+        assertEquals("Linked notification-base signature session response field 'sessionID' is missing or empty", ex.getMessage());
+    }
+}